moves subsidary funding out of token governance directory
