+++ src/components/DaoCreator/provider/types/index.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from 'ethers';
-import { NFTToFund, TokenToFund } from '../../tokenVotingGovernance/SubsidiaryFunding/types/index';
+import { NFTToFund, TokenToFund } from '../../SubsidiaryFunding/types/index';
 import { TokenAllocation } from '../../../../types/tokenAllocation';
 
 export enum CreatorProviderActions {
